Make routes tests use db mockup.
